visualizations_python/requirements.txt
kaleido==0.2.1
numpy==1.20.1
pandas==1.2.3
plotly==4.14.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
retrying==1.3.3
six==1.15.0
